cambios
creacion guia despacho en centro distribucion, con llamado a usuario y pedido, hay algunas dto que no estan en uso y que debo revisar.
crud incompleto, solo metodo crear guia esta listo, faltaria el resto de ese crud.
falta crud detalleguia
y verificar validaciones en los metodos

+++ ProyectoConMS/centro-distribucion-ms/src/main/java/cl/fullstack/centro_distribucion_ms/Client/PedidoClient.java
@@ -16,8 +16,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 public class PedidoClient {
 
     @Autowired
-    private RestTemplate restTemplate;  // ✅ Inyecta el bean con @LoadBalanced
-public List<PedidoDTO> obtenerPedidosPorDespachador(Long idDespachador) {
+    private RestTemplate restTemplate;  // 
+public List<PedidoDTO> obtenerPedidosPorDespachador(int idDespachador) {
     try {
         ResponseEntity<List<PedidoDTO>> response = restTemplate.exchange(
             "http://pedido-ms/api/pedidos/despachador/" + idDespachador,
@@ -27,7 +27,7 @@ public List<PedidoDTO> obtenerPedidosPorDespachador(Long idDespachador) {
         );
         return response.getBody();
     } catch (Exception e) {
-        e.printStackTrace();  // <--- agrega para ver qué error ocurre si falla
+        e.printStackTrace(); 
         return Collections.emptyList();
     }
 }
