+++ ProyectoConMS/centro-distribucion-ms/src/main/java/cl/fullstack/centro_distribucion_ms/dto/GuiaDespachoCreateRequest.java
@@ -4,5 +4,5 @@ import lombok.Data;
 
 @Data
 public class GuiaDespachoCreateRequest {
-    private Long idDespachador;
+    private int idDespachador;
 }
