+++ ProyectoConMS/centro-distribucion-ms/src/main/java/cl/fullstack/centro_distribucion_ms/repository/GuiaDespachoRepository.java
@@ -17,5 +17,5 @@ package cl.fullstack.centro_distribucion_ms.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import cl.fullstack.centro_distribucion_ms.entity.GuiaDespachoEntity;
 
-public interface GuiaDespachoRepository extends JpaRepository<GuiaDespachoEntity, Long> {
+public interface GuiaDespachoRepository extends JpaRepository<GuiaDespachoEntity, Integer> {
 }
